# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Book a meal is an app that enable users order their  favorite meal.Just by a single click, your meal is on your door step.",
    "version" : "0.2",
    "title" : "Book-A-Meal-API"
  },
  "host" : "bookmeals.herokuappcom",
...
  "paths" : {
...
    "/api/v1/orders" : {
      "post" : {
        "tags" : [ "Orders" ],
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "authorization",
          "in" : "header",
          "description" : "Generated header parameter. Example value ='eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjEsInVzZXJuYW1lIjoic3VuZGF5QSIsInJvbGUiOiJhZG1pbiIsImlhdCI6MTUyNjc1OTQ0OSwiZXhwIjoxNTI2ODQ1ODQ5fQ.g4x9TT1Lm7idUSMLhEB4tCt3MFCbM20wA2IT9qwhrZs'",
          "required" : false,
          "type" : "string",
          "x-example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjEsInVzZXJuYW1lIjoic3VuZGF5QSIsInJvbGUiOiJhZG1pbiIsImlhdCI6MTUyNjc1OTQ0OSwiZXhwIjoxNTI2ODQ1ODQ5fQ.g4x9TT1Lm7idUSMLhEB4tCt3MFCbM20wA2IT9qwhrZs"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model6"
          },
          "x-examples" : {
            "application/json" : "{\n    \"title\": \"today's menu\",\n    \"mealId\": 13,\n    \"menuId\": 4,\n    \"quantity\": 2,\n    \"address\": \"10 Reverend street Ojo\"\n}"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/postOrderModel"
            }
          }
        }
      },
      "get" : {
        "tags" : [ "Orders" ],
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "authorization",
          "in" : "header",
          "description" : "Generated header parameter. Example value ='eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjEsInVzZXJuYW1lIjoic3VuZGF5QSIsInJvbGUiOiJhZG1pbiIsImlhdCI6MTUyNjc1OTQ0OSwiZXhwIjoxNTI2ODQ1ODQ5fQ.g4x9TT1Lm7idUSMLhEB4tCt3MFCbM20wA2IT9qwhrZs'",
          "required" : false,
          "type" : "string",
          "x-example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjEsInVzZXJuYW1lIjoic3VuZGF5QSIsInJvbGUiOiJhZG1pbiIsImlhdCI6MTUyNjc1OTQ0OSwiZXhwIjoxNTI2ODQ1ODQ5fQ.g4x9TT1Lm7idUSMLhEB4tCt3MFCbM20wA2IT9qwhrZs"
        } ],
        "responses" : {
          "201" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/orderModel"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "postOrderModel" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "status" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "address" : {
          "type" : "string"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "totalPrice" : {
          "type" : "integer",
          "format" : "int32"
        },
        "catererId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "updatedAt" : {
          "type" : "string"
        },
        "createdAt" : {
          "type" : "string"
        },
        "userId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "meals" : {
          "type" : "Array"
        }
      },
      "example" : {
        "id" : 0,
        "catererId" : 5,
        "createdAt" : "createdAt",
        "address" : "address",
        "quantity" : 6,
        "meals" : [1, 2],
        "totalPrice" : 1,
        "title" : "title",
        "userId" : 5,
        "status" : "status",
        "updatedAt" : "updatedAt"
      }
    },
    "orderModel" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "status" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "address" : {
          "type" : "string"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "totalPrice" : {
          "type" : "integer",
          "format" : "int32"
        },
        "catererId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "updatedAt" : {
          "type" : "string"
        },
        "createdAt" : {
          "type" : "string"
        },
        "userId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "meals" : {
          "type" : "Array"
        }
      },
      "example" : [{
        "id" : 1,
        "catererId" : 5,
        "createdAt" : "createdAt",
        "address" : "address",
        "meals" : [{
          "createdAt" : "createdAt",
          "price" : 6,
          "name" : "name",
          "description" : "description",
          "id" : 0,
          "userId" : 1,
          "updatedAt" : "updatedAt",
          "orderMealItems": {
            "quantity": 3
          }
        },
        {
          "createdAt" : "createdAt",
          "price" : 6,
          "name" : "name",
          "description" : "description",
          "id" : 0,
          "userId" : 1,
          "updatedAt" : "updatedAt",
          "orderMealItems": {
            "quantity": 4
          }
        }],
        "totalPrice" : 12,
        "title" : "title",
        "userId" : 5,
        "status" : "pending",
        "updatedAt" : "updatedAt"
      }]
    },
...
    "Model6" : {
      "properties" : {
        "meals" : [{
          "type" : "integer",
          "format" : "int32",
          "mealId": 1,
          "menuId": 2,
          "quantity": 3
        },
        {
          "type" : "integer",
          "format" : "int32",
          "mealId": 2,
          "menuId": 3,
          "quantity": 4
        }],
        "address" : {
          "type" : "string"
        }
      },
      "example" : {
        "meals" : [{
          "mealId": 2,
          "menuId": 3,
          "quantity": 4
        },
        {"mealId": 1,
          "menuId": 2,
          "quantity": 3
        }
      ],
        "address" : "address"
      }
    },
...
  }
}